+++ src/constants/WalletAndCurrencyConstants.ts
@@ -356,7 +356,17 @@ export const SPECIAL_CURRENCY_INFO: {
     allowZeroTx: false,
     displayBuyCrypto: false,
     isImportKeySupported: true,
-    isStakingSupported: false,
+    isStakingSupported: true,
+    isCustomTokensSupported: false,
+    isPaymentProtocolSupported: false
+  },
+  filecoinfevmcalibration: {
+    initWalletName: lstrings.string_first_filecoin_fevm_calibratio_wallet_name,
+    chainCode: 'tFIL',
+    allowZeroTx: false,
+    displayBuyCrypto: false,
+    isImportKeySupported: true,
+    isStakingSupported: true,
     isCustomTokensSupported: false,
     isPaymentProtocolSupported: false
   },