+++ src/plugins/stake-plugins/types.ts
@@ -2,7 +2,7 @@
 // Stake Policy
 // -----------------------------------------------------------------------------
 
-import { EdgeAccount, EdgeCurrencyWallet } from 'edge-core-js'
+import { EdgeAccount, EdgeCorePluginOptions, EdgeCurrencyWallet } from 'edge-core-js'
 
 // -----------------------------------------------------------------------------
 // Errors
@@ -197,3 +197,5 @@ export interface StakePlugin {
   fetchChangeQuote: (request: ChangeQuoteRequest) => Promise<ChangeQuote>
   fetchStakePosition: (request: StakePositionRequest) => Promise<StakePosition>
 }
+
+export type StakePluginFactory = (opts?: EdgeCorePluginOptions) => Promise<StakePlugin>