+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - added: Add Pokt RPCs as option for Fantom
+- added: GLIF Infinity Pool staking plugin for Filecoin FEVM currency
 - added: Support for Thorchain Savers EVM token deposits
 - changed: USP/Legacy Landing probability set to 50/50
 - changed: Use memos for Thorchain Savers withdrawals